refactor: split file explorer tree item
